collector/epublica/requirements.txt
pyarrow 
boto3 
python-dotenv 
s3fs
pika
pandas
Minio
sentry_sdk==0.10.2
requests
pymongo